Implement multiple processes and threads to extract data

+++ core/extract/ExtractThread.js
@@ -1,16 +1,21 @@
-const {
-    Worker,
-    isMainThread,
-    parentPort,
-    workerData
-} = require("worker_threads");
+const {Worker} = require("worker_threads");
 
 class ExtractThread {
     constructor(params = {}) {
         this.modulePath = params.modulePath;
         this.socket = params.socket;
-        this.worker = null;
         this.status = false;
+        this.worker = new Worker(this.modulePath);
+        this.worker.on("online", () => (this.status = true));
+        this._bindListenEvent();
+        this.worker.postMessage({
+            type: "extract-start",
+            data: {catalogId: params.catalogId}
+        });
+        this.socket.emit("thread-info", {
+            processId: process.pid,
+            threadId: this.worker.threadId
+        });
     }
 
     getWorker() {
@@ -29,78 +34,62 @@ class ExtractThread {
         return this.status;
     }
 
-    initInstance() {
-        if (this.modulePath) {
-            this.worker = new Worker(this.modulePath);
-            this.worker.on("online", () => (this.status = true));
-            this._bindListenEvent();
-            this.socket.emit("thread-info", {
-                processId: process.pid,
-                threadId: this.worker.threadId
-            });
-        } else {
-            throw new Error("Module path is null!");
-        }
-    }
-
-    start(catalogId) {
-        this.worker.postMessage({
-            type: "start",
-            data: {catalogId: catalogId}
-        });
-    }
-
     pause() {
-        this.worker.postMessage({type: "pause"});
+        this.worker.postMessage({type: "extract-pause"});
     }
 
     continue() {
-        this.worker.postMessage({type: "continue"});
+        this.worker.postMessage({type: "extract-continue"});
     }
 
-    sendMessage(message) {
-        this.worker.send(message);
+    terminate() {
+        this.worker.postMessage({type: "extract-terminate"});
+        this.status = false;
     }
 
     _bindListenEvent() {
         this.worker.on("message", message => {
             switch (message.type) {
-                case "success":
+                case "extract-success":
                     this.socket.emit("thread-message", {
-                        type: "success",
+                        type: "extract-success",
                         message: `${message.data.url} - <b>${message.data.statusCode}</b>`
                     });
                     break;
-                case "error":
+                case "extract-error":
                     this.socket.emit("thread-message", {
-                        type: "error",
-                        message: `${JSON.stringify(message.data)}`
+                        type: "extract-error",
+                        message: `${message.data}`
                     });
                     break;
-                case "info":
+                case "extract-info":
                     if (message.data) {
                         this.socket.emit("thread-message", {
-                            type: "info",
+                            type: "extract-info",
                             message: message.data
                         });
                     }
                     break;
-                default:
+                case "extract-terminate":
+                    if (message.data) {
+                        this.worker.terminate();
+                    }
+                    break;
+                case "extract-finish":
+                    // finish
+                    this.socket.emit("thread-message", {
+                        type: "extract-info",
+                        message: `Extract data completed. Click <a href="#">here</a> to view.`
+                    });
                     break;
             }
         });
-
         this.worker.on("exit", () => {
             this.socket.emit("thread-message", {
                 type: "info",
                 message: `Thread was terminated...`
             });
         });
-    }
-
-    terminate() {
-        this.worker.terminate();
-        this.status = false;
     }
 }
